Add debug hooks to the event dispatcher, remove explicit prints
