+++ upy-freeze.txt
@@ -1,0 +1,1 @@
+vendor/upy-lib/logging/logging.py
